refactor: Page 컴포넌트를 atoms 디렉토리 하위로 이동
