+++ src/components/commons/index.js
@@ -1,9 +1,9 @@
 /* atoms */
 export { CartIcon } from './atoms/CartIcon';
 export { Line } from './atoms/Line';
+export { Page } from './atoms/Page';
 
 /* molecules */
 export { Header } from './molecules/Header';
 export { NavBar } from './molecules/NavBar';
-export { Page } from './molecules/Page';
 export { ProductItem } from './molecules/ProductItem';
